added migration ad-hoc files for apps in conf, and modify install script accordingly

+++ install.sh
@@ -12,6 +12,9 @@ usage : $0 --upgrade --dev --conf
     --upgrade | Upgrade the iskylims application
     --dev     | Use the develop version instead of main release
     --conf    | Select configuration file
+    --tables  | Load the first inital tables for upgrades
+    --script  | Run a migration script 
+    --ren_app | Rename applications required for the upgrade migration to 2.3.1 
 
 
 Examples:
@@ -30,7 +33,7 @@ EOF
 }
 
 db_check(){
-	mysqladmin -h $DB_SERVER_IP -u$DB_USER -p$DB_PASS -P$DB_PORT processlist >/tmp/null ###user should have mysql permission on remote server.
+    mysqladmin -h $DB_SERVER_IP -u$DB_USER -p$DB_PASS -P$DB_PORT processlist >/tmp/null ###user should have mysql permission on remote server.
 
     if ! [ $? -eq 0 ]; then
         echo -e "${RED}ERROR : Unable to connect to database. Check if your database is running and accessible${NC}"
@@ -106,6 +109,30 @@ root_check(){
     fi
 }
 
+update_settings_and_urls(){
+    # save SECRET KEY at home user directory
+    grep ^SECRET iSkyLIMS/settings.py > ~/.secret
+
+    # Copying config files and script. TODO CHANGE iSkyLIMS to app name
+    cp conf/template_settings.py $INSTALL_PATH/iSkyLIMS/settings.py
+    cp conf/urls.py $INSTALL_PATH/iSkyLIMS
+    
+    # replacing dummy variables with real values
+    sed -i "/^SECRET/c\\$(cat ~/.secret)" $INSTALL_PATH/iSkyLIMS/settings.py
+    sed -i "s/djangouser/${DB_USER}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    sed -i "s/djangopass/${DB_PASS}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    sed -i "s/djangohost/${DB_SERVER_IP}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    sed -i "s/djangoport/${DB_PORT}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    sed -i "s/djangodbname/${DB_NAME}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+
+    sed -i "s/emailhostserver/${EMAIL_HOST_SERVER}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    sed -i "s/emailport/${EMAIL_PORT}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    sed -i "s/emailhostuser/${EMAIL_HOST_USER}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    sed -i "s/emailhostpassword/${EMAIL_HOST_PASSWORD}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    sed -i "s/emailhosttls/${EMAIL_USE_TLS}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    sed -i "s/localserverip/${LOCAL_SERVER_IP}/g" $INSTALL_PATH/iSkyLIMS/settings.py    
+}
+
 #================================================================
 #SET TEMINAL COLORS
 #================================================================
@@ -127,12 +154,13 @@ do
     fi
     case "$arg" in
     ##OPTIONAL
-        --install)  set -- "$@"	-i ;;
-		--upgrade)	set -- "$@"	-u ;;
-        --script)   set -- "$@"	-s ;;
-        --tables)   set -- "$@"	-t ;;
+        --install)  set -- "$@" -i ;;
+        --upgrade)  set -- "$@" -u ;;
+        --script)   set -- "$@" -s ;;
+        --tables)   set -- "$@" -t ;;
         --dev)      set -- "$@" -d ;;
         --conf)     set -- "$@" -c ;;
+        --ren_app)  set -- "$@" -r ;;
 
     ## ADITIONAL
         --help)     set -- "$@" -h ;;
@@ -144,53 +172,57 @@ done
 
 #SETTING DEFAULT VALUES
 upgrade=false
+rename_applications=false
 git_branch="main"
 conf_file="./install_settings.txt"
 
 #PARSE VARIABLE ARGUMENTS WITH getops
-options=":c:dutsvh"
+options=":c:durtsvh"
 while getopts $options opt; do
-	case $opt in
+    case $opt in
         i ) type_installation=$OPTARG
             ;;
-		u )
-			upgrade=true
-			;;
+        u )
+            upgrade=true
+            ;;
         s )
-			run_script=true
+            run_script=true
             migration_script=$OPTARG
-			;;
+            ;;
         t )
-			update_tables=true
-			;;
-		d )
-			git_branch="develop"
-			;;
+            update_tables=true
+            ;;
+        r )
+            rename_applications=true
+            ;;
+        d )
+            git_branch="develop"
+            ;;
         c )
-		  	conf_file=$OPTARG
-		  	;;
-		h )
-		  	usage
-		  	exit 1
-		  	;;
-		v )
-		  	echo $ISKYLIMS_VERSION
-		  	exit 1
-		  	;;
-		\?)
-			echo "Invalid Option: -$OPTARG" 1>&2
-			usage
-			exit 1
-			;;
-		: )
-      		echo "Option -$OPTARG requires an argument." >&2
-      		exit 1
-      		;;
-      	* )
-			echo "Unimplemented option: -$OPTARG" >&2;
-			exit 1
-			;;
-	esac
+            conf_file=$OPTARG
+            ;;
+        h )
+            usage
+            exit 1
+            ;;
+        v )
+            echo $ISKYLIMS_VERSION
+            exit 1
+            ;;
+        \?)
+            echo "Invalid Option: -$OPTARG" 1>&2
+            usage
+            exit 1
+            ;;
+        : )
+            echo "Option -$OPTARG requires an argument." >&2
+            exit 1
+            ;;
+        * )
+            echo "Unimplemented option: -$OPTARG" >&2;
+            exit 1
+            ;;
+    esac
 done
 shift $((OPTIND-1))
 
@@ -251,11 +283,13 @@ if [ $upgrade == true ]; then
     printf "${YELLOW}------------------${NC}\n\n"
 
     ### Delete git and no copy files stuff
-    rm -rf $INSTALL_PATH/.git $INSTALL_PATH/.github $INSTALL_PATH/.gitignore $INSTALL_PATH/.Rhistory $INSTALL_PATH/docker-compose.yml $INSTALL_PATH/docker_iskylims_install.sh $INSTALL_PATH/Dockerfile $INSTALL_PATH/install.sh $INSTALL_PATH/install_settings.txt 
-    mv $INSTALL_PATH/iSkyLIMS_core $INSTALL_PATH/core 
-    mv $INSTALL_PATH/iSkyLIMS_wetlab $INSTALL_PATH/wetlab
-    mv $INSTALL_PATH/iSkyLIMS_drylab $INSTALL_PATH/drylab
-    mv $INSTALL_PATH/iSkyLIMS_clinic $INSTALL_PATH/clinic
+    if [ $rename_applications ] ; then
+        rm -rf $INSTALL_PATH/.git $INSTALL_PATH/.github $INSTALL_PATH/.gitignore $INSTALL_PATH/.Rhistory $INSTALL_PATH/docker-compose.yml $INSTALL_PATH/docker_iskylims_install.sh $INSTALL_PATH/Dockerfile $INSTALL_PATH/install.sh $INSTALL_PATH/install_settings.txt 
+        mv $INSTALL_PATH/iSkyLIMS_core $INSTALL_PATH/core 
+        mv $INSTALL_PATH/iSkyLIMS_wetlab $INSTALL_PATH/wetlab
+        mv $INSTALL_PATH/iSkyLIMS_drylab $INSTALL_PATH/drylab
+        mv $INSTALL_PATH/iSkyLIMS_clinic $INSTALL_PATH/clinic
+    fi
 
     # update installation by sinchronize folders
     echo "Copying files to installation folder"
@@ -267,51 +301,87 @@ if [ $upgrade == true ]; then
     echo "Installing required python packages"
     python3 -m pip install -r conf/requirements.txt
     
-    cp conf/template_settings.py $INSTALL_PATH/iSkyLIMS/settings.py
-    cp conf/urls.py $INSTALL_PATH/iSkyLIMS
+    # update the settings.py and the main urls
+    update_settings_and_urls
 
-    ## TODO grab secret from setting.py and put into .secret file in home
-    sed -i "/^SECRET/c\\$(cat ~/.secret)" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/djangouser/${DB_USER}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/djangopass/${DB_PASS}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/djangohost/${DB_SERVER_IP}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/djangoport/${DB_PORT}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/djangodbname/${DB_NAME}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    ### RENAME APP  in database and migration files ####
+    if [ $rename_applications ] ; then
+        # make migrations backup in home
+        # sed old app name to new app name to all migration scripts in migration folders. Always the app name and core in all
+        sed -i 's/iSkyLIMS_core/core/g' core/migrations/*.py
+        sed -i 's/iSkyLIMS_clinic/clinic/g' clinic/migrations/*.py
+        sed -i 's/iSkyLIMS_core/core/g' clinic/migrations/*.py
+        sed -i 's/iSkyLIMS_drylab/drylab/g' drylab/migrations/*.py
+        sed -i 's/iSkyLIMS_wetlab/wetlab/g' drylab/migrations/*.py
+        sed -i 's/iSkyLIMS_core/core/g' drylab/migrations/*.py
+        sed -i 's/iSkyLIMS_wetlab/wetlab/g' wetlab/migrations/*.py
+        sed -i 's/iSkyLIMS_drylab/drylab/g' wetlab/migrations/*.py
+        sed -i 's/iSkyLIMS_core/core/g' wetlab/migrations/*.py
+        sed -i 's/iSkyLIMS_core/core/g' core/migrations/*.py
 
-    sed -i "s/emailhostserver/${EMAIL_HOST_SERVER}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/emailport/${EMAIL_PORT}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/emailhostuser/${EMAIL_HOST_USER}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/emailhostpassword/${EMAIL_HOST_PASSWORD}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/emailhosttls/${EMAIL_USE_TLS}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/localserverip/${LOCAL_SERVER_IP}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+        mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP \
+            -e 'UPDATE django_content_type SET app_label = REPLACE(app_label , "iSkyLIMS_core", "core") WHERE app_label like ("iSkyLIMS_%");'
+        mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP  \
+            -e 'UPDATE django_content_type SET app_label = REPLACE(app_label , "iSkyLIMS_clinic", "clinic") WHERE app_label like ("iSkyLIMS_%");'
+        mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP \
+            -e 'UPDATE django_content_type SET app_label = REPLACE(app_label , "iSkyLIMS_wetlab", "wetlab") WHERE app_label like ("iSkyLIMS_%");'
+        mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP \
+            -e 'UPDATE django_content_type SET app_label = REPLACE(app_label , "iSkyLIMS_drylab", "drylab") WHERE app_label like ("iSkyLIMS_%");'
+        
+        mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP \
+            -e 'UPDATE django_migrations SET app = REPLACE(app , "iSkyLIMS_core", "core") WHERE app like ("iSkyLIMS_%");'
+        mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP \
+            -e 'UPDATE django_migrations SET app = REPLACE(app , "iSkyLIMS_clinic", "clinic") WHERE app like ("iSkyLIMS_%");'
+        mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP \
+            -e 'UPDATE django_migrations SET app = REPLACE(app , "iSkyLIMS_wetlab", "wetlab") WHERE app like ("iSkyLIMS_%");'
+        mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP \
+            -e 'UPDATE django_migrations SET app = REPLACE(app , "iSkyLIMS_drylab", "drylab") WHERE app like ("iSkyLIMS_%");'
+        
+        query_rename_table="SELECT CONCAT('RENAME TABLE ', TABLE_SCHEMA, '.', TABLE_NAME, \
+                            ' TO ', TABLE_SCHEMA, '.', REPLACE(TABLE_NAME, 'iSkyLIMS_', ''), ';') \
+                            AS query FROM information_schema.tables WHERE TABLE_SCHEMA = \"$DB_NAME\" AND TABLE_NAME LIKE 'iSkyLIMS_%';"
+        mysql -u $DB_USER -p$DB_PASS -h $DB_SERVER_IP -e "$query_rename_table" \
+            | xargs -I % echo "mysql -u$DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e \"% \" " | bash
 
-    ### RENAME APP ####
-    # make migrations backup in home
-    # sed old app name to new app name to all migration scripts in migration folders. Always the app name and core in all
-    sed -i 's/iSkyLIMS_clinic/clinic/g' clinic/migrations/*.py
-    sed -i 's/iSkyLIMS_core/core/g' clinic/migrations/*.py
-    sed -i 's/iSkyLIMS_drylab/drylab/g' drylab/migrations/*.py
-    sed -i 's/iSkyLIMS_wetlab/wetlab/g' drylab/migrations/*.py
-    sed -i 's/iSkyLIMS_core/core/g' drylab/migrations/*.py
-    sed -i 's/iSkyLIMS_wetlab/wetlab/g' wetlab/migrations/*.py
-    sed -i 's/iSkyLIMS_drylab/drylab/g' wetlab/migrations/*.py
-    sed -i 's/iSkyLIMS_core/core/g' wetlab/migrations/*.py
+        #query_rename_indexes="SELECT CONCAT('ALTER TABLE ', kcu.TABLE_SCHEMA, '.', kcu.TABLE_NAME, ' RENAME INDEX ', kcu.CONSTRAINT_NAME, \
+        #                       ' TO ', REPLACE(kcu.CONSTRAINT_NAME, 'iSkyLIMS_', ''), ';') \
+        #                       AS query FROM information_schema.key_column_usage kcu JOIN information_schema.table_constraints tc \
+        #                       ON tc.CONSTRAINT_NAME = kcu.CONSTRAINT_NAME WHERE kcu.TABLE_SCHEMA = \"$DB_NAME\" AND kcu.CONSTRAINT_NAME LIKE 'iSkyLIMS_%';"
+        #mysql -u $DB_USER -p$DB_PASS -h $DB_SERVER_IP -e "$query_rename_indexes"  \
+        #    | xargs -I % echo "mysql -u$DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e \"% \" " | bash
+        
+        query_rename_constraints="SELECT CONCAT('ALTER TABLE ', rcu.TABLE_SCHEMA, '.', rcu.TABLE_NAME, \
+                 ' DROP FOREIGN KEY ', rcu.CONSTRAINT_NAME, ';', \
+                 ' ALTER TABLE ', rcu.TABLE_SCHEMA, '.', rcu.TABLE_NAME, \
+                 ' ADD CONSTRAINT ', REPLACE(rcu.CONSTRAINT_NAME, 'iSkyLIMS_', ''), ' ', \
+                 tc.CONSTRAINT_TYPE, ' (', GROUP_CONCAT(rcu.COLUMN_NAME ORDER BY rcu.ORDINAL_POSITION SEPARATOR ', '), ')', \
+                 IF(tc.CONSTRAINT_TYPE = 'FOREIGN KEY', \
+                   CONCAT(' REFERENCES ', rcu.REFERENCED_TABLE_SCHEMA, '.', REPLACE(rcu.REFERENCED_TABLE_NAME, 'iSkyLIMS_', ''), ' (', \
+                          GROUP_CONCAT(rcu.REFERENCED_COLUMN_NAME ORDER BY rcu.ORDINAL_POSITION SEPARATOR ', '), ') ON DELETE ', rc.DELETE_RULE), \
+                   ''), ';') AS query \
+                FROM information_schema.key_column_usage rcu \
+                JOIN information_schema.table_constraints tc ON rcu.CONSTRAINT_NAME = tc.CONSTRAINT_NAME \
+                JOIN information_schema.referential_constraints rc ON rcu.CONSTRAINT_NAME = rc.CONSTRAINT_NAME \
+                WHERE rcu.TABLE_SCHEMA = '$DB_NAME' AND rcu.CONSTRAINT_NAME LIKE 'iSkyLIMS_%' \
+                GROUP BY rcu.TABLE_SCHEMA, rcu.TABLE_NAME, rcu.CONSTRAINT_NAME, tc.CONSTRAINT_TYPE, rcu.REFERENCED_TABLE_SCHEMA, rcu.REFERENCED_TABLE_NAME, rc.DELETE_RULE;"
+        mysql -u $DB_USER -p$DB_PASS -h $DB_SERVER_IP -e "$query_rename_constraints" | xargs -I % echo "mysql -u$DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e \"% \" " | bash
+        
+        # rm -rf ./*/migrations/__pycache__
+        # rm -rf ./*/migrations/__init__.py
+        
+        # copy modified migration files
+        cp conf/0002_core_migration_v2.3.1.py core/migrations
+        cp conf/0002_drylab_migration_v2.3.1.py drylab/migrations
+        cp conf/0002_wetlab_migration_v2.3.1.py wetlab/migrations
+        cp conf/0002_clinic_migration_v2.3.1.py clinic/migrations
+        cp conf/0002_django_utils_migration_v2.3.1.py django_utils/migrations
 
-    mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e 'UPDATE django_content_type SET app_label = REPLACE(app_label , "iSkyLIMS_core", "core") WHERE app_label like ("iSkyLIMS_%");'
-    mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e 'UPDATE django_content_type SET app_label = REPLACE(app_label , "iSkyLIMS_clinic", "clinic") WHERE app_label like ("iSkyLIMS_%");'
-    mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e 'UPDATE django_content_type SET app_label = REPLACE(app_label , "iSkyLIMS_wetlab", "wetlab") WHERE app_label like ("iSkyLIMS_%");'
-    mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e 'UPDATE django_content_type SET app_label = REPLACE(app_label , "iSkyLIMS_drylab", "drylab") WHERE app_label like ("iSkyLIMS_%");'
-    
-    mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e 'UPDATE django_migrations SET app = REPLACE(app , "iSkyLIMS_core", "core") WHERE app like ("iSkyLIMS_%");'
-    mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e 'UPDATE django_migrations SET app = REPLACE(app , "iSkyLIMS_clinic", "clinic") WHERE app like ("iSkyLIMS_%");'
-    mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e 'UPDATE django_migrations SET app = REPLACE(app , "iSkyLIMS_wetlab", "wetlab") WHERE app like ("iSkyLIMS_%");'
-    mysql -u $DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e 'UPDATE django_migrations SET app = REPLACE(app , "iSkyLIMS_drylab", "drylab") WHERE app like ("iSkyLIMS_%");'
-    
-    query="SELECT CONCAT('RENAME TABLE ', TABLE_SCHEMA, '.', TABLE_NAME, ' TO ', TABLE_SCHEMA, '.', REPLACE(TABLE_NAME, 'iSkyLIMS_', ''), ';') AS query FROM information_schema.tables WHERE TABLE_SCHEMA = \"$DB_NAME\" AND TABLE_NAME LIKE 'iSkyLIMS_%';"
-    mysql -u $DB_USER -p$DB_PASS -h $DB_SERVER_IP -e "$query" | xargs -I % echo "mysql -u$DB_USER -p$DB_PASS -D $DB_NAME -h $DB_SERVER_IP -e \"% \" " | bash
-
-    echo "checking for database changes"
-    ./manage.py makemigrations
+    else
+        echo "checking for database changes"
+        # rm -rf ./*/migrations/__pycache__
+        # rm -rf ./*/migrations/__init__.py
+        ./manage.py makemigrations
+    fi
     
     read -p "Do you want to proceed with the migrate command? (Y/N) " -n 1 -r
     echo    # (optional) move to a new line
@@ -559,31 +629,15 @@ if [ "$type_installation" = "full" ] || [ "$type_installation" = "application" ]
 
 
 
-
     # Starting iSkyLIMS
-        echo "activate the virtualenv"
-        source virtualenv/bin/activate
+    echo "activate the virtualenv"
+    source virtualenv/bin/activate
+
     echo "Creating iskylims project"
     django-admin startproject iSkyLIMS .
-    grep ^SECRET iSkyLIMS/settings.py > ~/.secret
-
-    # Copying config files and script. TODO CHANGE iSkyLIMS to app name
-    cp conf/template_settings.py $INSTALL_PATH/iSkyLIMS/settings.py
-    cp conf/urls.py $INSTALL_PATH/iSkyLIMS
-
-    sed -i "/^SECRET/c\\$(cat ~/.secret)" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/djangouser/${DB_USER}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/djangopass/${DB_PASS}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/djangohost/${DB_SERVER_IP}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/djangoport/${DB_PORT}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/djangodbname/${DB_NAME}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-
-    sed -i "s/emailhostserver/${EMAIL_HOST_SERVER}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/emailport/${EMAIL_PORT}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/emailhostuser/${EMAIL_HOST_USER}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/emailhostpassword/${EMAIL_HOST_PASSWORD}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/emailhosttls/${EMAIL_USE_TLS}/g" $INSTALL_PATH/iSkyLIMS/settings.py
-    sed -i "s/localserverip/${LOCAL_SERVER_IP}/g" $INSTALL_PATH/iSkyLIMS/settings.py
+    
+    # update the settings.py and the main urls
+    update_settings_and_urls
 
     echo "Creating the database structure for iSkyLIMS"
     python3 manage.py migrate
